Move WriteContext
